new: videorec api, screenshot_async() with fast PBOs, profile_enable()
new: demos/physics/ (tokamak),
new: ddraw_cube33(), ddraw_cylinder(), ddraw_prism(),
fix: rgba()/bgra(), new rgbaf()/bgraf(), cooker_progress() when cooking disabled, fix #21
fix: GLad API visibility from C++,
chg: non-variadic ddraw_text(),

ui: fix ui_bits() order,
3rd: rewrote glfw_gen.py (10 MiB smaller output),
3rd: glfw 3.4.0-wip -> 3.3.6-stable, lua 5.4.3 -> 5.4.4, miniaudio 0.10.42 -> 0.11.8, nuklear 4.04.1 -> 4.9.6,
vfs: reduce clutter in vfs_load() log,
file: remove double slashes in file_list() (osx),
misc: fix most compiler warnings (gcc/clang/tcc, linux/osx),
model: fix GL resource leak when rendering instanced models,
string: strbeg(), strend(),
luajit: prevent crash when binding lua on top of luajit (linux), do not version fwk.dll anymore,

+++ art/sources/fwk_ds.h
@@ -50,7 +50,7 @@ API size_t vlen( void* p );
 #define array_at(t,i) (t[i])
 #define array_count(t) (int)( (t) ? array_vlen_(t) / sizeof(0[t]) : 0u )
 #define array_bytes(t) (int)( (t) ? array_vlen_(t) : 0u )
-#define array_sort(t, cmpfunc) qsort( t, array_count(t), sizeof(0[t]), cmpfunc == strcmp ? strcmp_qsort : cmpfunc )
+#define array_sort(t, cmpfunc) qsort( t, array_count(t), sizeof(0[t]), (uintptr_t)cmpfunc == (uintptr_t)strcmp ? strcmp_qsort : cmpfunc )
 #define array_empty(t) ( !array_count(t) )
 static __thread unsigned array_c_;
